Ledger API Server H2 Database support (#2528)
