+++ ledger/sandbox/BUILD.bazel
@@ -94,6 +94,7 @@ da_scala_binary(
     visibility = ["//visibility:public"],
     deps = [
         ":sandbox",
+        "//3rdparty/jvm/com/h2database:h2",
     ],
 )
 
@@ -142,6 +143,7 @@ testDependencies = [
     "//3rdparty/jvm/org/scalacheck:scalacheck",
     "//3rdparty/jvm/org/awaitility:awaitility",
     "//3rdparty/jvm/commons_io:commons_io",
+    "//3rdparty/jvm/com/h2database:h2",
     "//bazel_tools/runfiles:scala_runfiles",
 ] + compileDependencies
 
@@ -251,6 +253,32 @@ client_server_test(
 ) if not is_windows else None
 
 client_server_test(
+    name = "conformance-test-static-time-h2database",
+    timeout = "long",
+    client = "//ledger/ledger-api-test-tool",
+    client_args = [
+        "--all-tests",
+        "--timeout-scale-factor=10",
+    ],
+    data = [
+        "//ledger/test-common:dar-files",
+    ],
+    server = "//ledger/sandbox:sandbox-binary",
+    server_args = [
+        "--port 0",
+        "--static-time",
+        "--eager-package-loading",
+        "--sql-backend-jdbcurl jdbc:h2:mem:static_time",
+    ],
+    server_files = [
+        "$(rootpaths //ledger/test-common:dar-files)",
+    ],
+    tags = [
+        "dont-run-on-darwin",
+    ],
+) if not is_windows else None
+
+client_server_test(
     name = "conformance-test-wall-clock-postgres",
     timeout = "long",
     client = "//ledger/ledger-api-test-tool:ledger-api-test-tool",
@@ -271,6 +299,33 @@ client_server_test(
         "--port 0",
         "--wall-clock-time",
         "--eager-package-loading",
+    ],
+    server_files = [
+        "$(rootpaths //ledger/test-common:dar-files)",
+    ],
+    tags = [
+        "dont-run-on-darwin",
+    ],
+) if not is_windows else None
+
+client_server_test(
+    name = "conformance-test-wall-clock-h2database",
+    timeout = "long",
+    client = "//ledger/ledger-api-test-tool:ledger-api-test-tool",
+    client_args = [
+        "--all-tests",
+        "--exclude TimeIT",
+        "--timeout-scale-factor=10",
+    ],
+    data = [
+        "//ledger/test-common:dar-files",
+    ],
+    server = "//ledger/sandbox:sandbox-binary",
+    server_args = [
+        "--port 0",
+        "--wall-clock-time",
+        "--eager-package-loading",
+        "--sql-backend-jdbcurl jdbc:h2:mem:wall_clock_time",
     ],
     server_files = [
         "$(rootpaths //ledger/test-common:dar-files)",
