+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/index/StandaloneIndexServer.scala
@@ -10,7 +10,6 @@ import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.Sink
 import com.daml.ledger.participant.state.v1.{ParticipantId, ReadService, WriteService}
-import com.digitalasset.api.util.TimeProvider
 import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.daml.lf.engine.Engine
 import com.digitalasset.grpc.adapter.ExecutionSequencerFactory
@@ -140,7 +139,7 @@ class StandaloneIndexServer(
 
     val initF = for {
       cond <- readService.getLedgerInitialConditions().runWith(Sink.head)
-      indexService <- PostgresIndex(
+      indexService <- JdbcIndex(
         readService,
         domain.LedgerId(cond.ledgerId),
         participantId,
@@ -162,7 +161,7 @@ class StandaloneIndexServer(
               writeService,
               indexService,
               StandaloneIndexServer.engine,
-              TimeProvider.UTC,
+              config.timeProvider,
               timeModel,
               SandboxConfig.defaultCommandConfig,
               None)(am, esf),
