+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/index/config/Config.scala
@@ -5,6 +5,7 @@ package com.digitalasset.platform.index.config
 
 import java.io.File
 
+import com.digitalasset.api.util.TimeProvider
 import com.digitalasset.ledger.api.tls.TlsConfiguration
 
 final case class Config(
@@ -12,6 +13,7 @@ final case class Config(
     portFile: Option[File],
     archiveFiles: List[File],
     maxInboundMessageSize: Int,
+    timeProvider: TimeProvider, // enables use of non-wall-clock time in tests
     jdbcUrl: String,
     tlsConfig: Option[TlsConfiguration]
 )
@@ -19,5 +21,5 @@ final case class Config(
 object Config {
   val DefaultMaxInboundMessageSize = 4194304
   def default: Config =
-    new Config(0, None, List.empty, 4194304, "", None)
+    new Config(0, None, List.empty, 4194304, TimeProvider.UTC, "", None)
 }
