+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/stores/ledger/sql/migration/FlywayMigrations.scala
@@ -3,20 +3,21 @@
 
 package com.digitalasset.platform.sandbox.stores.ledger.sql.migration
 
+import com.digitalasset.platform.sandbox.stores.ledger.sql.dao.JdbcLedgerDao
 import javax.sql.DataSource
 import org.flywaydb.core.Flyway
 import org.slf4j.LoggerFactory
 
 import scala.util.control.NonFatal
 
-class FlywayMigrations(ds: DataSource) {
+class FlywayMigrations(ds: DataSource, dbType: JdbcLedgerDao.DbType) {
   import FlywayMigrations._
 
   private val logger = LoggerFactory.getLogger(getClass)
 
   def migrate(): Unit = {
     try {
-      val flyway = configurationBase.dataSource(ds).load()
+      val flyway = configurationBase(dbType).dataSource(ds).load()
       logger.info(s"running Flyway migration..")
       val stepsTaken = flyway.migrate()
       logger.info(s"Flyway schema migration finished successfully applying ${stepsTaken} steps.")
@@ -33,7 +34,9 @@ class FlywayMigrations(ds: DataSource) {
 
 object FlywayMigrations {
 
-  val configurationBase = Flyway.configure()
+  def configurationBase(dbType: JdbcLedgerDao.DbType) =
+    Flyway.configure.locations("classpath:db/migration/" + dbType.name)
 
-  def apply(ds: DataSource): FlywayMigrations = new FlywayMigrations(ds)
+  def apply(ds: DataSource, dbType: JdbcLedgerDao.DbType): FlywayMigrations =
+    new FlywayMigrations(ds, dbType)
 }
