+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/stores/ledger/sql/migration/postgres/V3__Recompute_Key_Hash.scala
@@ -1,8 +1,9 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Note: package name must correspond exactly to the flyway 'locations' setting, which defaults to 'db.migration'
-package db.migration
+// Note: package name must correspond exactly to the flyway 'locations' setting, which defaults to
+// 'db.migration.postgres' for postgres migrations
+package db.migration.postgres
 
 import java.sql.{Connection, ResultSet}
